Implement evaluate on EvaluatableScalarRSQOneElectronOperator.
