+++ gqcp/tests/Operator/SecondQuantized/CMakeLists.txt
@@ -2,7 +2,6 @@ list(APPEND test_target_sources
     ${CMAKE_CURRENT_SOURCE_DIR}/EvaluatableScalarRSQOneElectronOperator_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/SimpleSQOneElectronOperator_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/SQHamiltonian_test.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/SQOneElectronOperator_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/SQTwoElectronOperator_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/USQHamiltonian_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/USQOneElectronOperator_test.cpp
